2008-09-23  Geoffrey Garen  <ggaren@apple.com>
        Try to fix the windows build.

        * VM/CTI.cpp:
        (JSC::CTI::compileOpCall):
        (JSC::CTI::privateCompileMainPass):



Canonical link: https://commits.webkit.org/29115@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@36841 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ JavaScriptCore/VM/CTI.cpp
@@ -512,8 +512,8 @@ void CTI::compileOpCall(Instruction* instruction, unsigned i, CompileOpCallType 
 
     // Initialize the parts of the call frame that have not already been initialized.
     emitGetCTIParam(CTI_ARGS_r, X86::edi);
-    m_jit.movl_i32m(reinterpret_cast<unsigned>(m_codeBlock), RegisterFile::CallerCodeBlock * sizeof(Register), X86::edi);
-    m_jit.movl_i32m(dst, RegisterFile::ReturnValueRegister * sizeof(Register), X86::edi);
+    m_jit.movl_i32m(reinterpret_cast<unsigned>(m_codeBlock), RegisterFile::CallerCodeBlock * static_cast<int>(sizeof(Register)), X86::edi);
+    m_jit.movl_i32m(dst, RegisterFile::ReturnValueRegister * static_cast<int>(sizeof(Register)), X86::edi);
 
     // Check the ctiCode has been generated - if not, this is handled in a slow case.
     m_jit.testl_rr(X86::eax, X86::eax);
@@ -676,7 +676,7 @@ void CTI::privateCompileMainPass()
 #if ENABLE(SAMPLING_TOOL)
             m_jit.movl_i32m(-1, &currentOpcodeID);
 #endif
-            m_jit.pushl_m(RegisterFile::ReturnPC * sizeof(Register), X86::edi);
+            m_jit.pushl_m(RegisterFile::ReturnPC * static_cast<int>(sizeof(Register)), X86::edi);
             m_jit.ret();
             i += 2;
             break;
@@ -973,7 +973,7 @@ void CTI::privateCompileMainPass()
         }
         case op_ret: {
             // Check for an activation - if there is one, jump to the hook below.
-            m_jit.cmpl_i32m(0, RegisterFile::OptionalCalleeActivation * sizeof(Register), X86::edi);
+            m_jit.cmpl_i32m(0, RegisterFile::OptionalCalleeActivation * static_cast<int>(sizeof(Register)), X86::edi);
             X86Assembler::JmpSrc activation = m_jit.emitUnlinkedJne();
             X86Assembler::JmpDst activated = m_jit.label();
 
@@ -990,11 +990,11 @@ void CTI::privateCompileMainPass()
             // Return the result in %eax, and the caller scope chain in %edx (this is read from the callee call frame,
             // but is only assigned to ExecState::m_scopeChain if returning to a JSFunction).
             emitGetArg(instruction[i + 1].u.operand, X86::eax);
-            m_jit.movl_mr(RegisterFile::CallerScopeChain * sizeof(Register), X86::edi, X86::edx);
+            m_jit.movl_mr(RegisterFile::CallerScopeChain * static_cast<int>(sizeof(Register)), X86::edi, X86::edx);
             // Restore the machine return addess from the callframe, roll the callframe back to the caller callframe,
             // and preserve a copy of r on the stack at CTI_ARGS_r. 
-            m_jit.movl_mr(RegisterFile::ReturnPC * sizeof(Register), X86::edi, X86::ecx);
-            m_jit.movl_mr(RegisterFile::CallerRegisters * sizeof(Register), X86::edi, X86::edi);
+            m_jit.movl_mr(RegisterFile::ReturnPC * static_cast<int>(sizeof(Register)), X86::edi, X86::ecx);
+            m_jit.movl_mr(RegisterFile::CallerRegisters * static_cast<int>(sizeof(Register)), X86::edi, X86::edi);
             emitPutCTIParam(X86::edi, CTI_ARGS_r);
 
             m_jit.pushl_r(X86::ecx);
